% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleior-package.R
\docType{package}
\name{pleior-package}
\alias{pleior}
\alias{pleior-package}
\title{pleior: Pleiotropy Analysis for GWAS Data}
\description{
The pleior package provides tools for identifying and analyzing pleiotropic genetic variants
in genome-wide association studies (GWAS). It includes functions for data loading,
preprocessing, pleiotropy detection, and visualization, enabling researchers to explore
shared genetic mechanisms across complex traits.
}
\references{
Watanabe, K., et al. (2019). A global overview of pleiotropy and genetic architecture
in complex traits. *Nature Genetics*, 51, 1339–1348.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danymukesha/pleior}
  \item \url{https://danymukesha.github.io/pleior/}
  \item Report bugs at \url{https://github.com/danymukesha/pleior/issues}
}

}
\author{
[Dany MUKESHA]
}
\keyword{GWAS}
\keyword{genetics}
\keyword{pleiotropy}
